+++ tests/check/check.go
@@ -1114,6 +1114,8 @@ type usersets-user
     define userset_to_butnot_computed: [directs-user#butnot_computed]
     define userset_to_and_computed:[directs-user#and_computed]
     define userset_recursive: [user, usersets-user#userset_recursive]
+    define userset_recursive_public: [user, user:*, usersets-user#userset_recursive_public]
+    define userset_recursive_public_only: [user:*, usersets-user#userset_recursive_public_only]
     define userset_recursive_mixed_direct_assignment: [user, usersets-user#userset_recursive_mixed_direct_assignment, usersets-user#userset]
     define or_userset: userset or userset_to_computed_cond
     define and_userset: userset_to_computed_cond and userset_to_computed_wild
@@ -1154,6 +1156,10 @@ type ttus
     define or_ttu: direct_pa_direct_ch or direct_cond_pa_direct_ch
     define and_ttu: or_comp_from_direct_parent and direct_pa_direct_ch
     define nested_butnot_ttu: or_comp_from_direct_parent but not userset_pa_userset_comp_wild_ch
+	define nested_ttu_parent: [ttus]
+	define nested_ttu: [directs-user] or nested_ttu from nested_ttu_parent
+	define nested_ttu_public: [directs-user, directs-user:*] or nested_ttu_public from nested_ttu_parent
+
 type complexity3
   relations
     define ttu_parent: [ttus]